+++ opentelemetry-sdk/tests/shared_internal/test_batch_processor.py
@@ -19,13 +19,10 @@ import os
 import time
 import unittest
 import weakref
-from concurrent.futures import ThreadPoolExecutor
 from platform import system
-from sys import version_info
 from unittest.mock import Mock
 
 import pytest
-from pytest import mark
 
 from opentelemetry.sdk._logs import (
     LogData,
@@ -34,6 +31,8 @@ from opentelemetry.sdk._logs import (
 from opentelemetry.sdk._logs.export import (
     BatchLogRecordProcessor,
 )
+from opentelemetry.sdk.trace import ReadableSpan
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
 from opentelemetry.sdk.util.instrumentation import InstrumentationScope
 
 EMPTY_LOG = LogData(
@@ -41,11 +40,21 @@ EMPTY_LOG = LogData(
     instrumentation_scope=InstrumentationScope("example", "example"),
 )
 
+BASIC_SPAN = ReadableSpan(
+    "MySpan",
+    instrumentation_scope=InstrumentationScope("example", "example"),
+)
 
-# BatchLogRecodpRocessor initializes / uses BatchProcessor.
+if system() != "Windows":
+    multiprocessing.set_start_method("fork")
+
+
+# BatchLogRecodProcessor/BatchSpanProcessor initialize and use BatchProcessor.
+# Important: make sure to call .shutdown() before the end of the test,
+# otherwise the worker thread will continue to run after the end of the test.
 @pytest.mark.parametrize(
     "batch_processor_class,telemetry",
-    [(BatchLogRecordProcessor, EMPTY_LOG)],
+    [(BatchLogRecordProcessor, EMPTY_LOG), (BatchSpanProcessor, BASIC_SPAN)],
 )
 class TestBatchProcessor:
     # pylint: disable=no-self-use
@@ -54,7 +63,7 @@ class TestBatchProcessor:
     ):
         exporter = Mock()
         batch_processor = batch_processor_class(
-            exporter=exporter,
+            exporter,
             max_queue_size=15,
             max_export_batch_size=15,
             # Will not reach this during the test, this sleep should be interrupted when batch size is reached.
@@ -63,13 +72,14 @@ class TestBatchProcessor:
         )
         before_export = time.time_ns()
         for _ in range(15):
-            batch_processor.emit(telemetry)
+            batch_processor._batch_processor.emit(telemetry)
         # Wait a bit for the worker thread to wake up and call export.
         time.sleep(0.1)
         exporter.export.assert_called_once()
         after_export = time.time_ns()
         # Shows the worker's 30 second sleep was interrupted within a second.
         assert after_export - before_export < 1e9
+        batch_processor.shutdown()
 
     # pylint: disable=no-self-use
     def test_telemetry_exported_once_schedule_delay_reached(
@@ -77,22 +87,23 @@ class TestBatchProcessor:
     ):
         exporter = Mock()
         batch_processor = batch_processor_class(
-            exporter=exporter,
+            exporter,
             max_queue_size=15,
             max_export_batch_size=15,
             schedule_delay_millis=100,
             export_timeout_millis=500,
         )
-        batch_processor.emit(telemetry)
+        batch_processor._batch_processor.emit(telemetry)
         time.sleep(0.2)
         exporter.export.assert_called_once_with([telemetry])
+        batch_processor.shutdown()
 
     def test_telemetry_flushed_before_shutdown_and_dropped_after_shutdown(
-        self, batch_processor_class, telemetry, caplog
+        self, batch_processor_class, telemetry
     ):
         exporter = Mock()
         batch_processor = batch_processor_class(
-            exporter=exporter,
+            exporter,
             # Neither of these thresholds should be hit before test ends.
             max_queue_size=15,
             max_export_batch_size=15,
@@ -100,15 +111,13 @@ class TestBatchProcessor:
             export_timeout_millis=500,
         )
         # This log should be flushed because it was written before shutdown.
-        batch_processor.emit(telemetry)
+        batch_processor._batch_processor.emit(telemetry)
         batch_processor.shutdown()
         exporter.export.assert_called_once_with([telemetry])
         assert batch_processor._batch_processor._shutdown is True
 
         # This should not be flushed.
-        batch_processor.emit(telemetry)
-        assert len(caplog.records) == 1
-        assert "Shutdown called, ignoring" in caplog.text
+        batch_processor._batch_processor.emit(telemetry)
         exporter.export.assert_called_once()
 
     # pylint: disable=no-self-use
@@ -117,7 +126,7 @@ class TestBatchProcessor:
     ):
         exporter = Mock()
         batch_processor = batch_processor_class(
-            exporter=exporter,
+            exporter,
             # Neither of these thresholds should be hit before test ends.
             max_queue_size=15,
             max_export_batch_size=15,
@@ -125,36 +134,10 @@ class TestBatchProcessor:
             export_timeout_millis=500,
         )
         for _ in range(10):
-            batch_processor.emit(telemetry)
+            batch_processor._batch_processor.emit(telemetry)
         batch_processor.force_flush()
         exporter.export.assert_called_once_with([telemetry for _ in range(10)])
-
-    @mark.skipif(
-        system() == "Windows" or version_info < (3, 9),
-        reason="This test randomly fails on windows and python 3.8.",
-    )
-    def test_with_multiple_threads(self, batch_processor_class, telemetry):
-        exporter = Mock()
-        batch_processor = batch_processor_class(
-            exporter=exporter,
-            max_queue_size=3000,
-            max_export_batch_size=1000,
-            schedule_delay_millis=30000,
-            export_timeout_millis=500,
-        )
-
-        def bulk_emit_and_flush(num_emit):
-            for _ in range(num_emit):
-                batch_processor.emit(telemetry)
-            batch_processor.force_flush()
-
-        with ThreadPoolExecutor(max_workers=69) as executor:
-            for idx in range(69):
-                executor.submit(bulk_emit_and_flush, idx + 1)
-
-            executor.shutdown()
-        # 69 calls to force flush.
-        assert exporter.export.call_count == 69
+        batch_processor.shutdown()
 
     @unittest.skipUnless(
         hasattr(os, "fork"),
@@ -175,13 +158,11 @@ class TestBatchProcessor:
         # _at_fork_reinit should be called in the child process, to
         # clear the logs/spans in the child process.
         for _ in range(9):
-            batch_processor.emit(telemetry)
-
-        multiprocessing.set_start_method("fork")
+            batch_processor._batch_processor.emit(telemetry)
 
         def child(conn):
             for _ in range(100):
-                batch_processor.emit(telemetry)
+                batch_processor._batch_processor.emit(telemetry)
             batch_processor.force_flush()
 
             # Expect force flush to export 10 batches of max export batch size (10)
@@ -196,6 +177,7 @@ class TestBatchProcessor:
         batch_processor.force_flush()
         # Single export for the telemetry we emitted at the start of the test.
         assert exporter.export.call_count == 1
+        batch_processor.shutdown()
 
     def test_record_processor_is_garbage_collected(
         self, batch_processor_class, telemetry