+++ opentelemetry-sdk/tests/trace/export/test_export.py
@@ -12,19 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import gc
-import multiprocessing
+
 import os
 import threading
 import time
 import unittest
-import weakref
-from concurrent.futures import ThreadPoolExecutor
-from logging import WARNING
-from platform import python_implementation, system
 from unittest import mock
-
-from pytest import mark
 
 from opentelemetry import trace as trace_api
 from opentelemetry.context import Context
@@ -35,13 +28,10 @@ from opentelemetry.sdk.environment_variables import (
     OTEL_BSP_MAX_QUEUE_SIZE,
     OTEL_BSP_SCHEDULE_DELAY,
 )
-from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import export
 from opentelemetry.sdk.trace.export import logger
-from opentelemetry.sdk.trace.export.in_memory_span_exporter import (
-    InMemorySpanExporter,
-)
-from opentelemetry.test.concurrency_test import ConcurrencyTestBase
+
+# pylint: disable=protected-access
 
 
 class MySpanExporter(export.SpanExporter):
@@ -154,23 +144,12 @@ class TestSimpleSpanProcessor(unittest.TestCase):
         self.assertListEqual([], spans_names_list)
 
 
-def _create_start_and_end_span(name, span_processor, resource):
-    span = trace._Span(
-        name,
-        trace_api.SpanContext(
-            0xDEADBEEF,
-            0xDEADBEEF,
-            is_remote=False,
-            trace_flags=trace_api.TraceFlags(trace_api.TraceFlags.SAMPLED),
-        ),
-        span_processor=span_processor,
-        resource=resource,
-    )
-    span.start()
-    span.end()
-
-
-class TestBatchSpanProcessor(ConcurrencyTestBase):
+# Many more test cases for the BatchSpanProcessor exist under
+# opentelemetry-sdk/tests/shared_internal/test_batch_processor.py.
+# Important: make sure to call .shutdown() on the BatchSpanProcessor
+# before the end of the test, otherwise the worker thread will continue
+# to run after the end of the test.
+class TestBatchSpanProcessor(unittest.TestCase):
     @mock.patch.dict(
         "os.environ",
         {
@@ -185,20 +164,38 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
             MySpanExporter(destination=[])
         )
 
-        self.assertEqual(batch_span_processor.max_queue_size, 10)
-        self.assertEqual(batch_span_processor.schedule_delay_millis, 2)
-        self.assertEqual(batch_span_processor.max_export_batch_size, 3)
-        self.assertEqual(batch_span_processor.export_timeout_millis, 4)
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_queue_size, 10
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._schedule_delay_millis, 2
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_export_batch_size, 3
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._export_timeout_millis, 4
+        )
+        batch_span_processor.shutdown()
 
     def test_args_env_var_defaults(self):
         batch_span_processor = export.BatchSpanProcessor(
             MySpanExporter(destination=[])
         )
 
-        self.assertEqual(batch_span_processor.max_queue_size, 2048)
-        self.assertEqual(batch_span_processor.schedule_delay_millis, 5000)
-        self.assertEqual(batch_span_processor.max_export_batch_size, 512)
-        self.assertEqual(batch_span_processor.export_timeout_millis, 30000)
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_queue_size, 2048
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._schedule_delay_millis, 5000
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_export_batch_size, 512
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._export_timeout_millis, 30000
+        )
+        batch_span_processor.shutdown()
 
     @mock.patch.dict(
         "os.environ",
@@ -216,10 +213,19 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
         )
         logger.disabled = False
 
-        self.assertEqual(batch_span_processor.max_queue_size, 2048)
-        self.assertEqual(batch_span_processor.schedule_delay_millis, 5000)
-        self.assertEqual(batch_span_processor.max_export_batch_size, 512)
-        self.assertEqual(batch_span_processor.export_timeout_millis, 30000)
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_queue_size, 2048
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._schedule_delay_millis, 5000
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._max_export_batch_size, 512
+        )
+        self.assertEqual(
+            batch_span_processor._batch_processor._export_timeout_millis, 30000
+        )
+        batch_span_processor.shutdown()
 
     def test_on_start_accepts_parent_context(self):
         # pylint: disable=no-self-use
@@ -237,149 +243,6 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
         span_processor.on_start.assert_called_once_with(
             span, parent_context=context
         )
-
-    def test_shutdown(self):
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(destination=spans_names_list)
-        span_processor = export.BatchSpanProcessor(my_exporter)
-
-        span_names = ["xxx", "bar", "foo"]
-
-        resource = Resource.create({})
-        for name in span_names:
-            _create_start_and_end_span(name, span_processor, resource)
-
-        span_processor.shutdown()
-        self.assertTrue(my_exporter.is_shutdown)
-
-        # check that spans are exported without an explicitly call to
-        # force_flush()
-        self.assertListEqual(span_names, spans_names_list)
-
-    def test_flush(self):
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(destination=spans_names_list)
-        span_processor = export.BatchSpanProcessor(my_exporter)
-
-        span_names0 = ["xxx", "bar", "foo"]
-        span_names1 = ["yyy", "baz", "fox"]
-
-        resource = Resource.create({})
-        for name in span_names0:
-            _create_start_and_end_span(name, span_processor, resource)
-
-        self.assertTrue(span_processor.force_flush())
-        self.assertListEqual(span_names0, spans_names_list)
-
-        # create some more spans to check that span processor still works
-        for name in span_names1:
-            _create_start_and_end_span(name, span_processor, resource)
-
-        self.assertTrue(span_processor.force_flush())
-        self.assertListEqual(span_names0 + span_names1, spans_names_list)
-
-        span_processor.shutdown()
-
-    def test_flush_empty(self):
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(destination=spans_names_list)
-        span_processor = export.BatchSpanProcessor(my_exporter)
-
-        self.assertTrue(span_processor.force_flush())
-
-    def test_flush_from_multiple_threads(self):
-        num_threads = 50
-        num_spans = 10
-
-        span_list = []
-
-        my_exporter = MySpanExporter(destination=span_list)
-        span_processor = export.BatchSpanProcessor(
-            my_exporter, max_queue_size=512, max_export_batch_size=128
-        )
-
-        resource = Resource.create({})
-
-        def create_spans_and_flush(tno: int):
-            for span_idx in range(num_spans):
-                _create_start_and_end_span(
-                    f"Span {tno}-{span_idx}", span_processor, resource
-                )
-            self.assertTrue(span_processor.force_flush())
-
-        with ThreadPoolExecutor(max_workers=num_threads) as executor:
-            future_list = []
-            for thread_no in range(num_threads):
-                future = executor.submit(create_spans_and_flush, thread_no)
-                future_list.append(future)
-
-            executor.shutdown()
-
-        self.assertEqual(num_threads * num_spans, len(span_list))
-
-    def test_flush_timeout(self):
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(
-            destination=spans_names_list, export_timeout_millis=500
-        )
-        span_processor = export.BatchSpanProcessor(my_exporter)
-
-        resource = Resource.create({})
-        _create_start_and_end_span("foo", span_processor, resource)
-
-        # check that the timeout is not meet
-        with self.assertLogs(level=WARNING):
-            self.assertFalse(span_processor.force_flush(100))
-        span_processor.shutdown()
-
-    def test_batch_span_processor_lossless(self):
-        """Test that no spans are lost when sending max_queue_size spans"""
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(
-            destination=spans_names_list, max_export_batch_size=128
-        )
-        span_processor = export.BatchSpanProcessor(
-            my_exporter, max_queue_size=512, max_export_batch_size=128
-        )
-
-        resource = Resource.create({})
-        for _ in range(512):
-            _create_start_and_end_span("foo", span_processor, resource)
-
-        time.sleep(1)
-        self.assertTrue(span_processor.force_flush())
-        self.assertEqual(len(spans_names_list), 512)
-        span_processor.shutdown()
-
-    def test_batch_span_processor_many_spans(self):
-        """Test that no spans are lost when sending many spans"""
-        spans_names_list = []
-
-        my_exporter = MySpanExporter(
-            destination=spans_names_list, max_export_batch_size=128
-        )
-        span_processor = export.BatchSpanProcessor(
-            my_exporter,
-            max_queue_size=256,
-            max_export_batch_size=64,
-            schedule_delay_millis=100,
-        )
-
-        resource = Resource.create({})
-        for _ in range(4):
-            for _ in range(256):
-                _create_start_and_end_span("foo", span_processor, resource)
-
-            time.sleep(0.1)  # give some time for the exporter to upload spans
-
-        self.assertTrue(span_processor.force_flush())
-        self.assertEqual(len(spans_names_list), 1024)
-        span_processor.shutdown()
 
     def test_batch_span_processor_not_sampled(self):
         tracer_provider = trace.TracerProvider(
@@ -402,134 +265,8 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
             pass
         time.sleep(0.05)  # give some time for the exporter to upload spans
 
-        self.assertTrue(span_processor.force_flush())
+        span_processor.force_flush()
         self.assertEqual(len(spans_names_list), 0)
-        span_processor.shutdown()
-
-    def _check_fork_trace(self, exporter, expected):
-        time.sleep(0.5)  # give some time for the exporter to upload spans
-        spans = exporter.get_finished_spans()
-        for span in spans:
-            self.assertIn(span.name, expected)
-
-    @unittest.skipUnless(
-        hasattr(os, "fork"),
-        "needs *nix",
-    )
-    def test_batch_span_processor_fork(self):
-        # pylint: disable=invalid-name
-        tracer_provider = trace.TracerProvider()
-        tracer = tracer_provider.get_tracer(__name__)
-
-        exporter = InMemorySpanExporter()
-        span_processor = export.BatchSpanProcessor(
-            exporter,
-            max_queue_size=256,
-            max_export_batch_size=64,
-            schedule_delay_millis=10,
-        )
-        tracer_provider.add_span_processor(span_processor)
-        with tracer.start_as_current_span("foo"):
-            pass
-        time.sleep(0.5)  # give some time for the exporter to upload spans
-
-        self.assertTrue(span_processor.force_flush())
-        self.assertEqual(len(exporter.get_finished_spans()), 1)
-        exporter.clear()
-
-        def child(conn):
-            def _target():
-                with tracer.start_as_current_span("span") as s:
-                    s.set_attribute("i", "1")
-                    with tracer.start_as_current_span("temp"):
-                        pass
-
-            self.run_with_many_threads(_target, 100)
-
-            time.sleep(0.5)
-
-            spans = exporter.get_finished_spans()
-            conn.send(len(spans) == 200)
-            conn.close()
-
-        parent_conn, child_conn = multiprocessing.Pipe()
-        p = multiprocessing.Process(target=child, args=(child_conn,))
-        p.start()
-        self.assertTrue(parent_conn.recv())
-        p.join()
-
-        span_processor.shutdown()
-
-    @mark.skipif(
-        python_implementation() == "PyPy" or system() == "Windows",
-        reason="This test randomly fails with huge delta in Windows or PyPy",
-    )
-    def test_batch_span_processor_scheduled_delay(self):
-        """Test that spans are exported each schedule_delay_millis"""
-        spans_names_list = []
-
-        export_event = threading.Event()
-        my_exporter = MySpanExporter(
-            destination=spans_names_list, export_event=export_event
-        )
-        start_time = time.time()
-        span_processor = export.BatchSpanProcessor(
-            my_exporter,
-            schedule_delay_millis=500,
-        )
-
-        # create single span
-        resource = Resource.create({})
-        _create_start_and_end_span("foo", span_processor, resource)
-
-        self.assertTrue(export_event.wait(2))
-        export_time = time.time()
-        self.assertEqual(len(spans_names_list), 1)
-        self.assertAlmostEqual((export_time - start_time) * 1e3, 500, delta=25)
-
-        span_processor.shutdown()
-
-    @mark.skipif(
-        python_implementation() == "PyPy" and system() == "Windows",
-        reason="This test randomly fails in Windows with PyPy",
-    )
-    def test_batch_span_processor_reset_timeout(self):
-        """Test that the scheduled timeout is reset on cycles without spans"""
-        spans_names_list = []
-
-        export_event = threading.Event()
-        my_exporter = MySpanExporter(
-            destination=spans_names_list,
-            export_event=export_event,
-            export_timeout_millis=50,
-        )
-
-        span_processor = export.BatchSpanProcessor(
-            my_exporter,
-            schedule_delay_millis=50,
-        )
-
-        with mock.patch.object(span_processor.condition, "wait") as mock_wait:
-            resource = Resource.create({})
-            _create_start_and_end_span("foo", span_processor, resource)
-            self.assertTrue(export_event.wait(2))
-
-            # give some time for exporter to loop
-            # since wait is mocked it should return immediately
-            time.sleep(0.1)
-            mock_wait_calls = list(mock_wait.mock_calls)
-
-            # find the index of the call that processed the singular span
-            for idx, wait_call in enumerate(mock_wait_calls):
-                _, args, __ = wait_call
-                if args[0] <= 0:
-                    after_calls = mock_wait_calls[idx + 1 :]
-                    break
-
-            self.assertTrue(
-                all(args[0] >= 0.05 for _, args, __ in after_calls)
-            )
-
         span_processor.shutdown()
 
     def test_batch_span_processor_parameters(self):
@@ -585,23 +322,6 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
             None,
             max_queue_size=256,
             max_export_batch_size=512,
-        )
-
-    def test_batch_span_processor_gc(self):
-        # Given a BatchSpanProcessor
-        exporter = MySpanExporter(destination=[])
-        processor = export.BatchSpanProcessor(exporter)
-        weak_ref = weakref.ref(processor)
-        processor.shutdown()
-
-        # When the processor is garbage collected
-        del processor
-        gc.collect()
-
-        # Then the reference to the processor should no longer exist
-        self.assertIsNone(
-            weak_ref(),
-            "The BatchSpanProcessor object created by this test wasn't garbage collected",
         )