+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/http_attributes.py
@@ -15,7 +15,7 @@
 from enum import Enum
 from typing import Final
 
-from deprecated import deprecated
+from typing_extensions import deprecated
 
 HTTP_CLIENT_IP: Final = "http.client_ip"
 """
@@ -74,7 +74,7 @@ The total size of the request in bytes. This should be the total number of bytes
 
 HTTP_REQUEST_CONTENT_LENGTH: Final = "http.request_content_length"
 """
-Deprecated: Replaced by `http.request.header.<key>`.
+Deprecated: Replaced by `http.request.header.content-length`.
 """
 
 HTTP_REQUEST_CONTENT_LENGTH_UNCOMPRESSED: Final = (
@@ -106,14 +106,14 @@ Deprecated in favor of stable :py:const:`opentelemetry.semconv.attributes.http_a
 
 HTTP_RESPONSE_CONTENT_LENGTH: Final = "http.response_content_length"
 """
-Deprecated: Replaced by `http.response.header.<key>`.
+Deprecated: hp.response.header.content-length.
 """
 
 HTTP_RESPONSE_CONTENT_LENGTH_UNCOMPRESSED: Final = (
     "http.response_content_length_uncompressed"
 )
 """
-Deprecated: Replace by `http.response.body.size`.
+Deprecated: Replaced by `http.response.body.size`.
 """
 
 HTTP_ROUTE: Final = "http.route"
@@ -123,7 +123,7 @@ Deprecated in favor of stable :py:const:`opentelemetry.semconv.attributes.http_a
 
 HTTP_SCHEME: Final = "http.scheme"
 """
-Deprecated: Replaced by `url.scheme` instead.
+Deprecated: Replaced by `url.scheme`.
 """
 
 HTTP_SERVER_NAME: Final = "http.server_name"
@@ -138,7 +138,7 @@ Deprecated: Replaced by `http.response.status_code`.
 
 HTTP_TARGET: Final = "http.target"
 """
-Deprecated: Split to `url.path` and `url.query.
+Deprecated: Split to `url.path` and `url.query`.
 """
 
 HTTP_URL: Final = "http.url"
@@ -160,8 +160,8 @@ class HttpConnectionStateValues(Enum):
 
 
 @deprecated(
-    reason="The attribute http.flavor is deprecated - Replaced by `network.protocol.name`"
-)  # type: ignore
+    "The attribute http.flavor is deprecated - Replaced by `network.protocol.name`"
+)
 class HttpFlavorValues(Enum):
     HTTP_1_0 = "1.0"
     """HTTP/1.0."""
@@ -178,8 +178,8 @@ class HttpFlavorValues(Enum):
 
 
 @deprecated(
-    reason="Deprecated in favor of stable :py:const:`opentelemetry.semconv.attributes.http_attributes.HttpRequestMethodValues`."
-)  # type: ignore
+    "Deprecated in favor of stable :py:const:`opentelemetry.semconv.attributes.http_attributes.HttpRequestMethodValues`."
+)
 class HttpRequestMethodValues(Enum):
     CONNECT = "CONNECT"
     """Deprecated in favor of stable :py:const:`opentelemetry.semconv.attributes.http_attributes.HttpRequestMethodValues.CONNECT`."""